Added Functionality to Save new Workouts to Database and Display Them in Table
Adding a workout will now save itself and all of its exercises into the database for that user. Needed to change the way in the dashboard the exercises were obtained from the user's custom workouts, but it works perfectly now. Just need to make it so the function is not just observing a single event, but constantly observing the database so it continuously updates the table. Also need to update the way the names and ID's of the new workouts are added to the table, as right now, for testing purposes, the ID's of the workouts are all 3, and the names of the children under the custom node are a random integer from 1 to 100 just to test that it was sucessfully adding multiple workouts as children in custom.